treat warning as error for base files (#3336)
* treat warning as error for base files

* remove UTFString as std::iterator template is deprecated in c++17

+++ native/cocos/base/Value.cpp
@@ -504,7 +504,7 @@ float Value::asFloat() const {
     }
 
     if (_type == Type::STRING) {
-        return utils::atof(_field.strVal->c_str());
+        return static_cast<float>(utils::atof(_field.strVal->c_str()));
     }
 
     if (_type == Type::INTEGER) {
